Refactor: Move flow postprocessors to their own package.
