+++ src/main/java/com/llamalad7/mixinextras/expression/impl/flow/FlowInterpreter.java
@@ -1,6 +1,8 @@
 package com.llamalad7.mixinextras.expression.impl.flow;
 
 import com.llamalad7.mixinextras.expression.impl.flow.expansion.ExpansionPostProcessor;
+import com.llamalad7.mixinextras.expression.impl.flow.postprocessing.FlowPostProcessor;
+import com.llamalad7.mixinextras.expression.impl.flow.postprocessing.NewArrayPostProcessor;
 import com.llamalad7.mixinextras.utils.TypeUtils;
 import org.objectweb.asm.Type;
 import org.objectweb.asm.tree.*;
