+++ src/main/java/com/llamalad7/mixinextras/expression/impl/flow/expansion/ExpansionPostProcessor.java
@@ -1,6 +1,6 @@
 package com.llamalad7.mixinextras.expression.impl.flow.expansion;
 
-import com.llamalad7.mixinextras.expression.impl.flow.FlowPostProcessor;
+import com.llamalad7.mixinextras.expression.impl.flow.postprocessing.FlowPostProcessor;
 import com.llamalad7.mixinextras.expression.impl.flow.FlowValue;
 
 import java.util.Arrays;
