+++ src/main/java/com/llamalad7/mixinextras/expression/impl/flow/expansion/IincExpander.java
@@ -1,6 +1,6 @@
 package com.llamalad7.mixinextras.expression.impl.flow.expansion;
 
-import com.llamalad7.mixinextras.expression.impl.flow.FlowPostProcessor;
+import com.llamalad7.mixinextras.expression.impl.flow.postprocessing.FlowPostProcessor;
 import com.llamalad7.mixinextras.expression.impl.flow.FlowValue;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.Type;
