+++ src/main/java/com/llamalad7/mixinextras/expression/impl/flow/postprocessing/ArrayCreationInfo.java
@@ -1,5 +1,6 @@
-package com.llamalad7.mixinextras.expression.impl.flow;
+package com.llamalad7.mixinextras.expression.impl.flow.postprocessing;
 
+import com.llamalad7.mixinextras.expression.impl.flow.FlowValue;
 import org.objectweb.asm.tree.AbstractInsnNode;
 
 import java.util.List;
